refactor(ControlsToggles): move stories inside component folder
