+++ src/components/ControlsToggles/stories/index.ts
@@ -1,0 +1,6 @@
+import './Checkbox.story';
+import './Radio.story';
+import './RadioGroup.story';
+import './Slider.story';
+import './Tabs.story';
+import './Toggle.story';
