+++ src/stories/index.ts
@@ -1,5 +1,5 @@
 import '@components/Buttons/stories';
-import './ControlsToggles.stories';
+import '@components/ControlsToggles/stories';
 import '@components/Icons/stories';
 import '@components/Inputs/stories';
 import '@components/Text/stories';
